requirements.txt
streamlit
pillow
pytesseract
language-tool-python
langdetect
transformers
torch
opencv-python-headless
numpy